My solution to week 66 challenge 1
Basic idea -- division is just repeated subtraction, so count how many times
we can subtract N from M.

If either M or N is negative, but not both, then the result will be negative.

If there's a remainder when the result is negative, then the floor is 1 more than
the count.