rework challenge-076-01 using goldbach's conjecture
looking through the other solutions i found this reference in @andinus
blog article:

https://stackoverflow.com/questions/35755825/find-minimum-prime-numbers-which-sum-to-a-given-value/35756072#35756072

I reworked my brute force solution to make use of the conjecture
assuming it is true for any N that will ever be feed to this programm.
The conjecture has been shown to hold for any $N < 4x10^18 which should
be sufficiently large.