Task #1: add more tests and fix the bugs uncovered
There were two problems uncovered by the tests, which include one test
that requires using a prime twice (6 = 3 + 3), one that is a prime (11),
and one that requires 3 terms (51 = 43 + 5 + 3 or 51 = 47 +2 + 2).

In find-min-primes-to-sum, I had the arguments to (conj) in the wrong
order for the case of requiring 3 terms.

In find-min-primes-to-sum-brute-force, I was not checking possibilities
involving the same primes twice, which I fixed by replacing
(combo/partitions) with (combo/selections) for both 2 and 3 terms.