Task 2: improve lonely detection
First, restrict search to elements that are X (or 1 in our case), so
we're not doing unnecessary work. Second, we can sum all the elements in
the submatrix and if they are equal to 1, then the only 1 is the element
we're checking and it is indeed surrounded by all 0s.