challenge-080 ch-1.pl: Fix logical error
It didn't check for =1=, I might have assumed that it was accounted
for in this line:
    print "1\n" and exit 0 if $sorted[$#sorted] < 1;

This was pointed out by <https://octodon.social/@polettix>
<https://tilde.zone/web/statuses/104981669595493301#>

I also discovered that it'll fail if the inputs are not unique, this
can be fixed by using hash & map somehow. Maybe I'll fix it later
someday.