Allow testing of SQL solutions.
Testing SQL requires to lot more work from the test program than testing
other languages. For other languages, they just read the input, and
calculate what ever needs to be calculated.

For SQL, things are a bit different. We need to create a database
(SQLite, in memory), create some tables (from a file .tables),
and, for each test, clear the tables, and fill it with data.
And since we have a habit to have actual multiple tests per file,
we need some machinery to deal with that setup.