Let the test file deal with place holder queries.
We're now dealing with queries with place holders. We're assuming
that input files will first give data for the tables, then a line
with values for the query.

For some reason, I can't get placeholders to work with SQLite and
SELECT queries, be it selectall_arrayref, or prepare/execute/
fetchall_arrayref. ("do" seems to handle place holders just fine).
Hence, we're manually replacing place holders in the query.

For now:
  - We're not quoting the value
  - We only deal with ? place holders, not ?1 style place holders.