Less restrictions on clues.
Some sudokus larger than 9x9 use all letters for clues. Some use
numbers 1 .. 9 followed by letters.

We're now excepting both. And we don't require them to be consecutive.

It's possible to have one less different clue that the size of the
sudoku, in that case, we have to make an educated guess (number or
letter) what the missing clue is.