Solve X-sudokus.
If the option --X is given, we're assuming the Sudoku is an X-Sudoku.
An X-Sudoku is a regular Sudoku with an additional constraint that
the number on the main and anti-diagonals are unique.

This was very easy to implement -- other than parsing the option,
all that needed to be changed was the method which determines what
other cells a particular cell can see.