Special case naked singles.
When we are solving, resolve naked singles without going into recursion.
Naked singles are cells which have only one possibility left.

For those cells, we cross off the possibility for all other unsolved
cells in the same row, column or box, and put the cell in the set
of solved cells. We keep resolving naked singles before recursing.