Rewrite of the Perl solution
This solution can be run direction, not requiring any options
(the previous solution required an undocumented -M5.010 option).
Instead of using atan2(), we now use a base-60 encoded list
of digts of pi (after the decimal dot).

The source code is 1485 characters, including white space.
The output is 1487 characters, of which 1485 digits (the others
are a decimal dot, and a new line).