Pointless splitting into sub strings.
The challenge asks to first split the given string into substring.
But for the calculation of the end result, we don't actually need
those substrings. To make sure we won't be disqualified, we split
the given strings into substrings -- which we tell ignore.

We only do this for the Perl solution; other language solutions will
not include pointless statements.