Don't rely on buggy Befunge::Interpreter
We're using a patched version of Befunge::Interpreter fixing some
of its glaring bugs. But the Befunge-93 specification says that
a space should be printed after printing a number.
Befunge::Interpreter does not do that, and we had not compensated
for this bug yet. But that made our tests succeed, when they
should fail.

So, now, instead of printing a single number, we now split the
number in tens and units, and print the number as characters.

While doing this, we found yet another Befunge::Interpreter which
needed to be compensated for.