A solution for week 125, part 2, for some definition of solution.
Once again, the problem does not specify how to binary tree input
is going to look like. All we can deduce from the single example
input is that is absolutly won't scale beyond utter trivial sizes.

So, we will just die on any input which isn't exactly the one case
which is specified.