Challenge-236 - "Exact Change" and "Count Loops"
Exact Change - use a simple algorithm to manage the wallet.
The problem description makes it seem like there might be a need
to recursively try multiple solutions, but in practice this is
not necessary.

Count Loops - I originally implemented this using a destructive
approach with a hash, but in my benchmarks this was slow and also
non-determinisitc, and making it deterministic would make it slower.
Using an array based approach seems like it would be worse, but
in practice, with randomly generated input data a pure array
approach was faster. *Shrug*